Fix use of size_t in printf formats.  Patch #5727.

+++ src/data/por-file-reader.c
@@ -430,9 +430,9 @@ read_version_data (struct pfm_reader *r, struct pfm_read_info *info)
 
   /* Validate file. */
   if (strlen (date) != 8)
-    error (r, _("Bad date string length %d."), strlen (date));
+    error (r, _("Bad date string length %d."), (int) strlen (date));
   if (strlen (time) != 6)
-    error (r, _("Bad time string length %d."), strlen (time));
+    error (r, _("Bad time string length %d."), (int) strlen (time));
 
   /* Save file info. */
   if (info != NULL)
